Ensure the formsets attribute keeps its order
`NestedMeta#formsets` can now be defined as a list of two-tuples,
holding the name and the formset class. The resulting
`NestedModelForm#formsets` will be an OrderedDict.

This change is backwards compatible.  `NestedMeta#formsets` still
supports being defined as a `dict`, but the ordering of formsets in this
case is undefined, just as it was previously.